Add auth handlers as router
