+++ apps/api/src/routes/index.ts
@@ -1,0 +1,1 @@
+export { default as authRouter } from "./authRouter";
